refactor(code): phase 3.4 code audit — extract shared config and hooks
- Add app/config/navigation.ts: single source of truth for nav links
- Add app/config/social.ts: single source of truth for social links + types
- Add app/hooks/use-reduced-motion.ts: shared hook (was duplicated in 2 components)
- header.tsx, mobile-menu.tsx, footer.tsx: consume NAVIGATION_LINKS from config
- footer.tsx, social-icons.tsx: consume SOCIAL_LINKS + types from config
- hero-animations.tsx, testimonials-carousel.tsx: use useReducedMotion hook
- hero.tsx: remove export from HERO_VARIANTS/HERO_BREAKPOINTS (internal only)
- hero-animations.tsx: remove unused FadeInContainer export
- content.server.ts: remove getYogaContent (never called)
- lib/index.ts: remove PREGNANCY_PERFORMANCE/PERFORMANCE_THRESHOLDS from barrel
- newsletter-input.tsx: fix unused catch variable (_err)
