+++ app/components/layout/event-card/event-card.tsx
@@ -1,6 +1,6 @@
 import type * as React from "react";
-import { cn } from "~/lib/utils";
 import { Button } from "~/components/ui/button";
+import { cn } from "~/lib/utils";
 
 function normalizeMonthName(monthName: string) {
   return monthName
