+++ app/components/layout/event-card/index.ts
@@ -1,2 +1,2 @@
-export { EventCard } from "./event-card";
 export type { EventCardProps } from "./event-card";
+export { EventCard } from "./event-card";
