+++ app/components/layout/header/index.ts
@@ -1,5 +1,4 @@
-export { Header } from "./header";
-export { MobileMenu } from "./mobile-menu";
-
 export type { HeaderProps } from "./header";
+export { Header } from "./header";
 export type { MobileMenuProps } from "./mobile-menu";
+export { MobileMenu } from "./mobile-menu";
