+++ app/components/layout/testimonials-carousel/index.ts
@@ -1,4 +1,4 @@
-export { TestimonialsCarousel } from "./testimonials-carousel";
-export type { TestimonialsCarouselProps, Testimonial } from "./testimonials-carousel";
-export { TestimonialCard } from "./testimonial-card";
 export type { TestimonialCardProps } from "./testimonial-card";
+export { TestimonialCard } from "./testimonial-card";
+export type { Testimonial, TestimonialsCarouselProps } from "./testimonials-carousel";
+export { TestimonialsCarousel } from "./testimonials-carousel";
