+++ app/config/navigation.ts
@@ -1,0 +1,7 @@
+/** Shared navigation links used by Header, MobileMenu, and Footer. */
+export const NAVIGATION_LINKS = [
+  { label: "Doula", href: "/doula" },
+  { label: "Yoga", href: "/yoga" },
+  { label: "Féminin", href: "/feminin-sacre" },
+  { label: "À propos", href: "/a-propos" },
+] as const;
