+++ app/lib/index.ts
@@ -2,38 +2,25 @@
  * Shooting Star Library Exports
  *
  * Centralized exports for all library utilities and functionality.
- * Import utilities from this file for consistent API access.
  *
  * @example
  * ```tsx
- * import { cn, prefersReducedMotion, PERFORMANCE_THRESHOLDS } from '~/lib';
+ * import { cn, prefersReducedMotion, sanitizeInput } from '~/lib';
  * ```
  */
 
 // Browser Support Detection (motion preferences only)
 export { prefersReducedMotion } from "~/lib/browser-support";
-
-// CSS Utilities
-export { cn } from "~/lib/utils";
+// Contact Form Schema (shared between client and server)
+export {
+  type ContactFormData,
+  contactFormSchema,
+} from "~/lib/contact-form-schema";
 
 // Form Security Utilities
 export { sanitizeInput } from "~/lib/form-security";
-
-// Contact Form Schema (shared between client and server)
-export {
-  contactFormSchema,
-  type ContactFormData,
-} from "~/lib/contact-form-schema";
+// CSS Utilities
+export { cn } from "~/lib/utils";
 
 // Rate Limiting — import directly from ~/lib/rate-limiter.server
 // (server-only module, not re-exported to prevent client bundle leakage)
-
-// Performance Thresholds
-export {
-  CORE_WEB_VITALS,
-  PREGNANCY_PERFORMANCE,
-  PERFORMANCE_THRESHOLDS,
-  type CoreWebVitals,
-  type PregnancyPerformance,
-  type PerformanceThresholds,
-} from "~/lib/performance-thresholds";
